Add support for derived time series
Derived time series are time series that are not stored. Their values
are instead computed by applying a function to the values of another
time series. This source time series must currently be a stored time
series, and the function must be pure (i.e. it cannot store any state
and should not have any side-effects). A derived time must be defined
as a triple consisting of the name of the source series, the name of
the derived series, and the definition of the function to apply.
Alternatively, the derived time series can be defined as the id of the
source series, the name of the derived series, and the definition of
the function to apply. The function to apply is called for each value
of the source time series and has access to said value and the scaling
factor for the derived time series using the symbols value and
scalingFactor, respectively. Currently, scalingFactor is always one
for derived time series.