Normalize predicate push-down for Derby and H2
Derby now emits a warning, instead of terminating the query with an
exception if the predicate cannot be pushed to storage, as many queries
otherwise could not run.

Derby now returns an iterator that wraps the ResultSet instead of an
iterator from a list so segment groups can be streamed from storage in
multiple batches.

H2 and Derby now exposes all SQL statements used to create the views and
UDAFs as variables and methods the companion object, allowing the views
to include columns defined at runtime and statements used in tests.

Replaced // with /* Documention blocks in all of H2's classes if there
are multiple links to documentation.