Add support for H2 as a query engine for edge nodes
Support for H2 as a query engine and storage layer has been added as a
more resource-efficient alternative to Apache Spark and Apache
Cassandra. Thus, ModelarDB can use H2 on the edge due to its lower
resource requirement, and Apache Spark and Apache Cassandra in the cloud
and data centers where scalability is required. As H2 is a full RDBMS
the SQLite JDBC driver is no longer bundled with ModelarDB but can be
added to the JVM classpath at run-time.

Additionally, multiple minor improvements have been implemented: All
existing dependencies have been updated, and IndexedRDD has been
vendored as it must be manually compiled for Scala 2.12. Support for
ingesting time series from ORC and Parquet files have been added. The
terminology used in the implementation has been normalized and updated
to match the latest papers. A default time zone is now set to ensure the
same time zone is used everywhere. All UDFs and UDAFs in the Segment
View now support #, and the _SS version of the UDAFs have been removed
as they are not need for H2 and can be replaced by INLINE for Apache
Spark. Last, validation of the error bound has been moved from
Configuration to the model types as the interpretation of the error
bound can differ between model types.

Co-authored-by: Soeren Kejser Jensen <devel@kejserjensen.dk>
Co-authored-by: Christian <5160123+c5m@users.noreply.github.com>